addition of obstacles and comments

+++ src/Edge.java
@@ -1,8 +1,10 @@
 
 class Edge
 {
+	//represents edges between verties
     public final Vertex target;
-    public final double weight;
+    public final int weight;
     public Edge(Vertex argTarget)
+    //edge weight is 1 
     { target = argTarget; weight = 1; }
 }
